Add more CMA metadata to log messages
Add `asyncOperationId`, `granules` (list of `granuleId`s), `parentArn`,
and `stackName` to all log messages.

Also, modified unit tests using `assertTrue` for equality testing to
use `assertEqual` instead so that a failing test will show the actual
and expected values rather than the unhelpful message "False is not true"

Also fixed bug where disabling the message adapter via setting the
environment variable `CUMULUS_MESSAGE_ADAPTER_DISABLED` to `"true"` has
no affect due to the following identity test (where
`message_adapter_disabled` is set to the value of the env var):

    if message_adapter_disabled is 'true':

While the unit test for disabling the message adapter passes, it does
not accurately reflect the behavior of externally setting the env var,
in which case `is` cannot be used to check for value equality.  That is,
the literal `true` above will not be the identical string read from the
environment.

CUMULUS-1635